requirements.txt
streamlit>=1.35.0
anthropic>=0.26.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
